% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rough_convert.R
\name{add_svgs_to_ggplot}
\alias{add_svgs_to_ggplot}
\title{Add svg to chart_container object}
\usage{
add_svgs_to_ggplot(chart_container)
}
\arguments{
\item{chart_container}{A chart_container}
}
\value{
A chart_container with svg key added
}
\description{
This function add a `svg` key containing a `svg` version of the chart in the
chart container.
}
