% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rough_convert.R
\name{add_svgs_to_table}
\alias{add_svgs_to_table}
\title{Add svg to chart_container object}
\usage{
add_svgs_to_table(chart_container, width = NULL, height = NULL)
}
\arguments{
\item{chart_container}{A chart_container}
}
\value{
A chart_container with svg key added
}
\description{
This function add a `svg` key containing a `svg` version of the chart in the
chart container. It uses the width and height of RStudio Viewer panel. There
are no API to grab this information, so we use `grDevices::dev.size` which
should be in sync with the Viewer panel. We use 0.9 to add a 10% margin.
See: https://stackoverflow.com/a/41401158/2008527
}
