% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rough_convert.R
\name{get_node_attrs}
\alias{get_node_attrs}
\title{Convert XML node attributes (and optionally content) to a list}
\usage{
get_node_attrs(nodes, add_content = F)
}
\arguments{
\item{nodes}{an `xml2` `nodeset`}

\item{add_content}{a boolean}
}
\value{
A list of character vectors
}
\description{
This function extends `xml2::xml_attrs()`, by adding the option to add the
node content as an additional key `content` to the list created by
`xml2::xml_attrs()`.
}
