% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rough_convert.R
\name{get_rough_chart}
\alias{get_rough_chart}
\title{Create a roughjs chart}
\usage{
get_rough_chart(p, rough_user_options = NULL, width = NULL, height = NULL,
  family = NULL, font_size_booster = 1)
}
\description{
Create a roughjs chart
}
