% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rough_convert.R
\name{keys_check}
\alias{keys_check}
\title{Test if keys are in list}
\usage{
keys_check(l, req_keys)
}
\arguments{
\item{l}{a list}

\item{req_keys}{a character vector of required keys}
}
\value{
A boolean, TRUE if the required keys req_keys are all in list l
}
\description{
This function takes a list and a vector of required keys. It returns TRUE
if all the required keys are in the list. Additional keys are allowed.
}
