% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rough_convert.R
\name{nv_to_list}
\alias{nv_to_list}
\title{Convert a named vector to a named list}
\usage{
nv_to_list(nc)
}
\arguments{
\item{nc}{A named vector}
}
\value{
A list with keys matching nc names
}
\description{
This function takes a named vector and convert it to a list with keys named
after the vector's names.
}
